refactor(cli): extract session runner
