+++ packages/agent-cli/agent-cli.cabal
@@ -59,6 +59,7 @@ library
                     , Agent.CLI.Session.Interaction
                     , Agent.CLI.Session.Lifecycle
                     , Agent.CLI.Session.Retry
+                    , Agent.CLI.Session.Runner
                     , Agent.CLI.Session.Selection
                     , Agent.CLI.Session.Runtime.Types
                     , Agent.CLI.SessionAdmin
